aider: Added favicon and improved typography.
# Aider chat conversation:

ASSISTANT: Sure, let's start with the first suggestion: Adding a favicon.

1. **Add a Favicon**

You need to add a link to your favicon in the head section of your HTML file. The favicon should be a .ico file and it's usually stored in the root directory of your website.

```html
index.html
<<<<<<< HEAD
    <title>Animated Matrix</title>
    <link
      href="https://fonts.googleapis.com/css2?family=JetBrains+Mono&amp;display=swap"
      rel="stylesheet"
    />
=======
    <title>Animated Matrix</title>
    <link rel="icon" href="favicon.ico" type="image/x-icon">
    <link
      href="https://fonts.googleapis.com/css2?family=JetBrains+Mono&amp;display=swap"
      rel="stylesheet"
    />
>>>>>>> updated
```

Please replace "favicon.ico" with the path to your actual favicon file.

2. **Improve Typography**

Let's change the font of the headings to make them stand out. We'll use Google Fonts for this. Add this link in your head section to import the 'Roboto' font.

```html
index.html
<<<<<<< HEAD
    <link
      href="https://fonts.googleapis.com/css2?family=JetBrains+Mono&amp;display=swap"
      rel="stylesheet"
    />
=======
    <link
      href="https://fonts.googleapis.com/css2?family=JetBrains+Mono|Roboto:wght@500&amp;display=swap"
      rel="stylesheet"
    />
>>>>>>> updated
```

Then, update your CSS to apply the 'Roboto' font to all headings (h1, h2, h3, etc.).

```html
index.html
<<<<<<< HEAD
    <style>
      body {
        font-family: "JetBrains Mono", monospace;
        background-color: #000;
        color: #0f0;
        margin: 0;
        overflow: hidden;
      }
=======
    <style>
      body {
        font-family: "JetBrains Mono", monospace;
        background-color: #000;
        color: #0f0;
        margin: 0;
        overflow: hidden;
      }
      h1, h2, h3, h4, h5, h6 {
        font-family: 'Roboto', sans-serif;
      }
>>>>>>> updated
```

Please note that these changes are just the beginning. I'll continue with the rest of the improvements in the next messages.