aider: Replace empty `src` attributes with valid image URLs in `gallery.html`.
# Aider chat conversation:

USER: add `gallery.html`

ASSISTANT: Sure, I can help with that. However, I need more specific instructions on what changes you want to make to the `gallery.html` file. Could you please provide more details?

USER: Fix the images not displaying correctly

ASSISTANT: The issue seems to be with the `img` tags that have empty `src` attributes. These tags are not pointing to any image, hence no image is displayed.

To fix this, we need to replace the empty `src` attributes with valid image URLs.

Here is the edit block for the changes:

```html
gallery.html
<<<<<<< HEAD
    <img src="" onclick="openPopup(this)" data-description="Image Description"><button class="remove-button" data-url="">x</button>
    <img src="" onclick="openPopup(this)" data-description="Image Description"><button class="remove-button" data-url="">x</button>
=======
    <img src="https://example.com/image1.jpg" onclick="openPopup(this)" data-description="Image Description"><button class="remove-button" data-url="https://example.com/image1.jpg">x</button>
    <img src="https://example.com/image2.jpg" onclick="openPopup(this)" data-description="Image Description"><button class="remove-button" data-url="https://example.com/image2.jpg">x</button>
>>>>>>> updated
```

Please replace `https://example.com/image1.jpg` and `https://example.com/image2.jpg` with the actual URLs of the images you want to display.

USER: InvalidEditBlock: edit failed!

gallery.html does not contain the *exact sequence* of HEAD lines you specified.
Try again.
DO NOT skip blank lines, comments, docstrings, etc!
The HEAD block needs to be EXACTLY the same as the lines in gallery.html with nothing missing!

gallery.html does not contain these 2 exact lines in a row:
```
    <img src="" onclick="openPopup(this)" data-description="Image Description"><button class="remove-button" data-url="">x</button>
    <img src="" onclick="openPopup(this)" data-description="Image Description"><button class="remove-button" data-url="">x</button>
```

ASSISTANT: I apologize for the confusion. It seems there was a misunderstanding in identifying the exact lines to be replaced. Let's try again.

The images that are not displaying correctly are the ones with empty `src` attributes. We need to replace these empty `src` attributes with valid image URLs.

Here is the correct edit block for the changes:

```html
gallery.html
<<<<<<< HEAD
    <img src="" onclick="openPopup(this)" data-description="Image Description"><button class="remove-button" data-url="">x</button>
=======
    <img src="https://example.com/image1.jpg" onclick="openPopup(this)" data-description="Image Description"><button class="remove-button" data-url="https://example.com/image1.jpg">x</button>
>>>>>>> updated
```

```html
gallery.html
<<<<<<< HEAD
    <img src="" onclick="openPopup(this)" data-description="Image Description"><button class="remove-button" data-url="">x</button>
=======
    <img src="https://example.com/image2.jpg" onclick="openPopup(this)" data-description="Image Description"><button class="remove-button" data-url="https://example.com/image2.jpg">x</button>
>>>>>>> updated
```

Please replace `https://example.com/image1.jpg` and `https://example.com/image2.jpg` with the actual URLs of the images you want to display.

USER: InvalidEditBlock: edit failed!

gallery.html does not contain the *exact sequence* of HEAD lines you specified.
Try again.
DO NOT skip blank lines, comments, docstrings, etc!
The HEAD block needs to be EXACTLY the same as the lines in gallery.html with nothing missing!

gallery.html does not contain these 1 exact lines in a row:
```
    <img src="" onclick="openPopup(this)" data-description="Image Description"><button class="remove-button" data-url="">x</button>
```

ASSISTANT: I apologize for the confusion. It seems there was a misunderstanding in identifying the exact lines to be replaced. Let's try again.

The images that are not displaying correctly are the ones with empty `src` attributes. We need to replace these empty `src` attributes with valid image URLs.

Here is the correct edit block for the changes:

```html
gallery.html
<<<<<<< HEAD
    <img src="" onclick="openPopup(this)" data-description="Image Description"><button class="remove-button" data-url="">x</button><img src="" onclick="openPopup(this)" data-description="Image Description"><button class="remove-button" data-url="">x</button><img src="https://pbs.twimg.com/media/F1Dr4CTXoAAP3AW?format=jpg&name=orig" onclick="openPopup(this)" data-description="Image Description"><button class="remove-button" data-url="https://pbs.twimg.com/media/F1Dr4CTXoAAP3AW?format=jpg&name=orig">x</button></div>
=======
    <img src="https://example.com/image1.jpg" onclick="openPopup(this)" data-description="Image Description"><button class="remove-button" data-url="https://example.com/image1.jpg">x</button><img src="https://example.com/image2.jpg" onclick="openPopup(this)" data-description="Image Description"><button class="remove-button" data-url="https://example.com/image2.jpg">x</button><img src="https://pbs.twimg.com/media/F1Dr4CTXoAAP3AW?format=jpg&name=orig" onclick="openPopup(this)" data-description="Image Description"><button class="remove-button" data-url="https://pbs.twimg.com/media/F1Dr4CTXoAAP3AW?format=jpg&name=orig">x</button></div>
>>>>>>> updated
```

Please replace `https://example.com/image1.jpg` and `https://example.com/image2.jpg` with the actual URLs of the images you want to display.