aider: Change the height of the images in the gallery to a fixed value of 200px.
# Aider chat conversation:

USER: make the display of the images more square right now they are verticcal

ASSISTANT: To make the images display more square, we can adjust the CSS for the images in the `gallery.html` file. Specifically, we can modify the height of the images in the `.image-grid img` selector. Currently, it's set to `25vw` which means the height of the images is 25% of the viewport width. This can make the images appear vertical if the viewport width is much larger than its height.

We can change this to a fixed pixel value or a percentage of the viewport height (`vh`) to make the images more square. Let's try setting it to `200px` for a start.

Here is the *edit block* for this change:

```html
gallery.html
<<<<<<< HEAD
    .image-grid img {
      width: 100%;
      cursor: pointer;
      object-fit: cover;
      height: 25vw; /* Set a fixed height for the grid items */
    }
=======
    .image-grid img {
      width: 100%;
      cursor: pointer;
      object-fit: cover;
      height: 200px; /* Set a fixed height for the grid items */
    }
>>>>>>> updated
```

Please note that the actual pixel value may need to be adjusted based on the specific requirements of your project and the dimensions of the images you are working with.