aider: Add a sidebar on the left to the gallery page to navigate to home.html and make it look elegant.
# Aider chat conversation:

USER: Add a side bar on the left to the gallery page to navigate to home.html and make it look elegant

ASSISTANT: To add a sidebar to the gallery page, we need to modify the `gallery.html` file. We will add a new `div` for the sidebar at the top of the `body` section, and then add a link to `home.html` inside this `div`. We will also need to add some CSS to style the sidebar and make it look elegant.

Here is the *edit block* for the HTML changes:

```html
gallery.html
<<<<<<< HEAD
<body>
  <div class="container">
=======
<body>
  <div class="sidebar">
    <a href="home.html">Home</a>
  </div>
  <div class="container">
>>>>>>> updated
```

And here is the *edit block* for the CSS changes:

```html
gallery.html
<<<<<<< HEAD
    .upload-button:hover {
      background-color: #007B9A; /* Darker blue */
    }
    .gallery-header {
      font-family: 'Fira Code', monospace;
      font-style: italic;
      font-size: 24px;
      text-align: center;
    }
=======
    .upload-button:hover {
      background-color: #007B9A; /* Darker blue */
    }
    .sidebar {
      position: fixed;
      height: 100%;
      width: 0;
      top: 0;
      left: 0;
      background-color: #111;
      overflow-x: hidden;
      transition: 0.5s;
      padding-top: 60px;
    }
    .sidebar a {
      padding: 8px 8px 8px 32px;
      text-decoration: none;
      font-size: 25px;
      color: #818181;
      display: block;
      transition: 0.3s;
    }
    .sidebar a:hover {
      color: #f1f1f1;
    }
    .gallery-header {
      font-family: 'Fira Code', monospace;
      font-style: italic;
      font-size: 24px;
      text-align: center;
      margin-left: 50px;
    }
>>>>>>> updated
```

This will create a sidebar that is initially hidden and slides out when the mouse hovers over it. The link to `home.html` is styled to be large and easy to click, and it changes color when the mouse hovers over it. The gallery header is also moved to the right to make room for the sidebar.