aider: Set up lazy loading for images.
# Aider chat conversation:

USER: set up lazy loading for images

ASSISTANT: To implement lazy loading for images, we will need to modify the `ai_images.html` file. We will use the `loading="lazy"` attribute for the `img` tag. This attribute is a native way to defer the loading of off-screen images and iframes until users scroll near them.

Here is the step-by-step explanation of the changes:

1. In the JavaScript code where we create the `img` element, we will add a new line to set the `loading` attribute of the `img` element to `"lazy"`.

Here is the *edit block* for the change:

```html
ai_images.html
<<<<<<< HEAD
                    const img = document.createElement('img');
                    img.src = image.src;
                    img.alt = image.alt;
=======
                    const img = document.createElement('img');
                    img.src = image.src;
                    img.alt = image.alt;
                    img.loading = "lazy";
>>>>>>> updated
```

This change will make the browser load the images only when they are about to scroll into the viewport, which can help improve the performance of the page, especially if there are many images or if the images are large.