aider: Created a script that automatically updates the `image_store.json` file based on the contents of the `imgs` directory.
# Aider chat conversation:

USER: is it possible to update the `image_store.json` based on what files are in the imgs folder automatically. If possible create a new script that tracks the changes and updates the json automatically.

ASSISTANT: Yes, it is possible to create a script that automatically updates the `image_store.json` file based on the contents of the `imgs` directory. We can use Python for this task. The script will:

1. List all files in the `imgs` directory.
2. Filter out non-image and non-video files.
3. Create a JSON object for each file with the necessary properties.
4. Write the JSON objects to the `image_store.json` file.

Here is the new Python script file that will do this:

```python
new_file.py
<<<<<<< HEAD
=======
import os
import json

# Directory containing images and videos
directory = './imgs'

# Get all files in the directory
files = os.listdir(directory)

# Filter out non-image and non-video files
media_files = [f for f in files if f.endswith(('.png', '.jpg', '.jpeg', '.mp4'))]

# Create a JSON object for each file
images = []
for file in media_files:
    src = f'./imgs/{file}'
    alt = f'Output {file}'
    description = f'Description for {file}'
    type = 'video' if file.endswith('.mp4') else 'image'
    images.append({'src': src, 'alt': alt, 'description': description, 'type': type})

# Write the JSON objects to the image_store.json file
with open('image_store.json', 'w') as f:
    json.dump({'images': images}, f, indent=2)
>>>>>>> updated
```

You can run this script whenever you want to update the `image_store.json` file. If you want to automate this process, you can use a task scheduler like cron (on Unix-based systems) or Task Scheduler (on Windows) to run the script at regular intervals.